set underlaydirbase in docwiki.setup
4291f2e3d7d1eb160427764e622d247145652e81 made
just setting underlaydir not enough to change
where other underlays are loaded from. While generally
a good thing, that broke building of the docwiki
from the source tree, because underlaydirbase
was not pointing there.

Setting underlaydirbase in the setup file is a solution,
albeit slightly hacky (since the variable is undocumented and internal).