remove wikilink to ikiwiki/directive/shortcut
This is a broken link if ikiwiki is configured w/o
the listdirectives plugin. Since the value of the
link to the directive docs is fairly low (there are
a bunch of examples of it right on the page
you put it on), it seems simplest to just remove the link.