remove link to page outside the base wiki
I'm sorry, but you can't simply go around adding links to pages that are
part of the basewiki, that point to pages outside the basewiki. Doing
so causes the test suite to fail, since the basewiki is then not
self-contained and has broken links.

FYI, the ikiwiki/anchor page you added is not part of the basewiki because
there's not a symlink to it in underlays/basewiki/ikiwiki in ikiwiki's
source. And I can't condone cluttering up the basewiki with such a page,
at least unless it gets significantly more useful somehow.