switch plugins page to use a map instead of an inline
A map just seems a better fit for reference, since it deliniates the
contrib plugins better, and orders better.

It also has the advantage of being less expensive, since the plugins page
does not need to update when eg, the pagecount page changes.

Only downside is, no rss feed of new plugins. Which I know a few people
were subscribed to.