Space sorting orders consistently
Markdown interprets the empty lines resulting from [[!if]] as a request
that each <li> contain a <p>. For consistent spacing in the HTML, either
the unconditional items should have blank lines between (resulting in
each <li> having a <p>), or the conditional items should not (resulting
in messy source code, and no <li> having a <p>). I think the former
looks nicer.