improved fix for depends_simple_mixup
Avoid adding the page matched against as an influence for
currently failing pagespec matches, while still adding
any other influences.

This avoids bloating depends_simple with lots of bogus influences when
matching eg, "!link(done)". It's only necessary for the page being tested
to be an influence of that if the page matches.