use html5 semantic markup in page.tmpl
This is a first pass, it avoids needing to change style.css
except where it refers to tag types.

This goes a bit off the rails at the pageheader with its nested header.
Semantically, there should be an article around the whole page
header, content, and footer. Just as there will be an article around a
whole comment or inlined page header, content, and footer.

But that will mean changing the css that currently refers to pageheader to
refer to the enclosing article instead.