adapt to ikiwiki
Upstream ships a collection of icons, but the licences of them are very
unclear, since most seem to be taken from the various openid provider
websites. That can't be included in ikiwiki. So, instead hotlink to
favicons of sites, and for large display, include the site name.

Removed vidoop.com, which is gone.

If an url is passed to init as the second parameter, add a "Local Login"
provider, which just links to do=signin.